-Integrate DeviceManager into RuntimeAPI
