+++ src/test/kotlin/Test.kt
@@ -1,7 +1,6 @@
 import kuda.driverapi.DriverAPI
 import kuda.driverapi.prop.Result
 import kuda.kublas.Kublas
-import kuda.runtimeapi.DeviceManager
 import kuda.runtimeapi.RuntimeAPI
 import kuda.runtimeapi.prop.FunctionCache
 import kuda.runtimeapi.prop.Limit
@@ -22,9 +21,9 @@ class Test {
         val cudaVersion = runtimeAPI.runtimeGetVersion()
         val driverVersion = driverAPI.driverGetVersion()
 
-        val device = DeviceManager.getDevice()
+        val device = runtimeAPI.getDevice()
 
-        DeviceManager.initDevice(device, 0)
+        runtimeAPI.initDevice(device, 0)
 
         val driverDevice = driverAPI.deviceGet(0)
         val deviceCount = driverAPI.deviceGetCount()
@@ -34,20 +33,20 @@ class Test {
         println("CUDA Device: $device")
         println("CUDA Device count : $deviceCount")
         println(driverDevice)
-        println(DeviceManager.getLimit(Limit.PRINT_FIFO_SIZE))
-        println(DeviceManager.getLimit(Limit.MALLOC_HEAP_SIZE))
-        println(DeviceManager.getLimit(Limit.STACK_SIZE))
-        println(DeviceManager.getLimit(Limit.DEV_RUNTIME_SYNC_DEPTH))
-        println(DeviceManager.getLimit(Limit.MAX_L2_FETCH_GRANULARITY))
-        println(DeviceManager.getLimit(Limit.DEV_RUNTIME_PENDING_LAUNCH_COUNT))
-        println(DeviceManager.getLimit(Limit.PERSISTING_L2_CACHE_SIZE))
+        println(runtimeAPI.getLimit(Limit.PRINT_FIFO_SIZE))
+        println(runtimeAPI.getLimit(Limit.MALLOC_HEAP_SIZE))
+        println(runtimeAPI.getLimit(Limit.STACK_SIZE))
+        println(runtimeAPI.getLimit(Limit.DEV_RUNTIME_SYNC_DEPTH))
+        println(runtimeAPI.getLimit(Limit.MAX_L2_FETCH_GRANULARITY))
+        println(runtimeAPI.getLimit(Limit.DEV_RUNTIME_PENDING_LAUNCH_COUNT))
+        println(runtimeAPI.getLimit(Limit.PERSISTING_L2_CACHE_SIZE))
 
-        println(DeviceManager.getPCIBusId(device))
-        println(DeviceManager.getStreamPriorityRange())
+        println(runtimeAPI.getPCIBusId(device))
+        println(runtimeAPI.getStreamPriorityRange())
 
-        println(DeviceManager.setCacheConfig(FunctionCache.PREFER_NONE))
-        println(DeviceManager.synchronize())
-        println(DeviceManager.reset())
+        println(runtimeAPI.setCacheConfig(FunctionCache.PREFER_NONE))
+        println(runtimeAPI.synchronize())
+        println(runtimeAPI.reset())
 
         println(runtimeAPI.driverGetVersion())
 
@@ -63,7 +62,7 @@ class Test {
         println(mallocHostPointer)
         runtimeAPI.freeHost(mallocHostPointer)
 
-        DeviceManager.getMemPool(device)
+        runtimeAPI.getMemPool(device)
 
         var eventStart =  runtimeAPI.eventCreate()
         var eventEnd = runtimeAPI.eventCreate()
@@ -80,18 +79,20 @@ class Test {
 
     @Test
     fun `test getDeviceProperties`(){
-        val device = DeviceManager.getDevice()
-        DeviceManager.initDevice(device, 0)
-        val prop = DeviceManager.getDeviceProperties(device)
+        val runtimeAPI = RuntimeAPI()
+        val device = runtimeAPI.getDevice()
+        runtimeAPI.initDevice(device, 0)
+        val prop = runtimeAPI.getDeviceProperties(device)
     }
 
     @Test
     fun `test chooseDevice`(){
-        val device = DeviceManager.getDevice()
-        DeviceManager.initDevice(device, 0)
-        val prop = DeviceManager.getDeviceProperties(device)
+        val runtimeAPI = RuntimeAPI()
+        val device = runtimeAPI.getDevice()
+        runtimeAPI.initDevice(device, 0)
+        val prop = runtimeAPI.getDeviceProperties(device)
 
-        println(DeviceManager.chooseDevice(prop))
+        println(runtimeAPI.chooseDevice(prop))
     }
 
     @Test
